+++ weblate/trans/forms.py
@@ -30,6 +30,7 @@ from crispy_forms.bootstrap import TabHolder, Tab, InlineRadios
 
 from django import forms
 from django.core.exceptions import PermissionDenied
+from django.core.validators import MaxValueValidator, MinValueValidator
 from django.utils.translation import (
     ugettext_lazy as _, ugettext, pgettext_lazy, pgettext
 )
@@ -52,6 +53,7 @@ from weblate.trans.models.source import PRIORITY_CHOICES
 from weblate.trans.models.unit import (
     STATE_TRANSLATED, STATE_FUZZY, STATE_APPROVED, STATE_EMPTY,
 )
+from weblate.trans.machine import MACHINE_TRANSLATION_SERVICES
 from weblate.permissions.helpers import (
     can_author_translation, can_overwrite_translation, can_translate,
     can_suggest, can_add_translation, can_mass_add_translation, can_review,
@@ -898,10 +900,29 @@ class AutoForm(forms.Form):
         required=False,
         initial=False
     )
+    auto_source = forms.ChoiceField(
+        label=_('Automatic translation source'),
+        choices=[
+            ('others', _('Other translation components')),
+            ('mt', _('Machine translation')),
+        ],
+        initial='others',
+    )
     subproject = forms.ChoiceField(
         label=_('Component to use'),
         required=False,
         initial=''
+    )
+    engines = forms.MultipleChoiceField(
+        label=_('Machine translation engines to use'),
+        choices=[],
+        required=False,
+    )
+    threshold = forms.IntegerField(
+        label=_("Score threshold"),
+        initial=80,
+        min_value=1,
+        max_value=100,
     )
 
     def __init__(self, obj, user, *args, **kwargs):
@@ -926,6 +947,27 @@ class AutoForm(forms.Form):
 
         self.fields['subproject'].choices = \
             [('', _('All components in current project'))] + choices
+        self.fields['engines'].choices = [
+            (key, mt.name) for key, mt in MACHINE_TRANSLATION_SERVICES.items()
+        ]
+        if 'weblate' in MACHINE_TRANSLATION_SERVICES.keys():
+            self.fields['engines'].initial = 'weblate'
+
+        self.helper = FormHelper(self)
+        self.helper.layout = Layout(
+            Field('overwrite'),
+            Field('inconsistent'),
+            InlineRadios('auto_source', id='select_auto_source'),
+            Div(
+                'subproject',
+                css_id='auto_source_others'
+            ),
+            Div(
+                'engines',
+                'threshold',
+                css_id='auto_source_mt'
+            ),
+        )
 
 
 class CommaSeparatedIntegerField(forms.Field):